+++ doc/source/whats_new.rst
@@ -11,6 +11,18 @@ Changelog
 
    - Linear (and zeroth-order) detrending for Epochs and Evoked by `Eric Larson`_
 
+   - Label morphing between subjects by `Eric Larson`_
+
+   - Define events based on time lag between reference and target event by `Denis Engemann`_
+
+   - ICA convenience function implementing an automated artifact removal workflow by `Denis Engemann`_
+
+   - Bad channels no longer included in epochs by default by `Eric Larson`_
+
+
+API
+~~~
+   - Deprecated use of fiff.pick_types without specifying exclude -- use either [] (none), 'bads' (bad channels), or a list of string (channel names).
 
 .. _changes_0_5: